fix(dispatch): restore core detect helper and CI gates
Ship pane-dispatch-detect.ps1 with core support so core/security
profiles keep the helpers winsmux-core.ps1 already calls. Extract
shell_prompt_ready from operator_cli.rs to return under the v0.36.26
physical-line baseline, and mock launch capture-pane so the in-process
send test can observe the typed fragment.
