+++ install.ps1
@@ -242,6 +242,9 @@ function Write-InstallProfileManifest {
 
 function Install-CoreSupportScripts {
     Download-File "winsmux-core/scripts/json-compat.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "json-compat.ps1")
+    # winsmux-core.ps1 calls Test-PaneContainsCommandFragment / Test-ShellPromptText
+    # on every profile, including core and security.
+    Download-File "winsmux-core/scripts/pane-dispatch-detect.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-dispatch-detect.ps1")
     Download-OptionalFile "winsmux-core/scripts/control-plane-workers.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "control-plane-workers.ps1")
     Download-OptionalFile "winsmux-core/scripts/control-plane-ledger.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "control-plane-ledger.ps1")
 }
@@ -280,7 +283,6 @@ function Install-OrchestraSupportScripts {
     Download-File "winsmux-core/scripts/orchestra-layout.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "orchestra-layout.ps1")
     Download-File "winsmux-core/scripts/orchestra-ui-attach.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "orchestra-ui-attach.ps1")
     Download-File "winsmux-core/scripts/pane-control.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-control.ps1")
-    Download-File "winsmux-core/scripts/pane-dispatch-detect.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-dispatch-detect.ps1")
     Download-File "winsmux-core/scripts/pane-env.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-env.ps1")
     Download-File "winsmux-core/scripts/pane-border.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-border.ps1")
     Download-File "winsmux-core/scripts/pane-status.ps1" (Join-Path $BRIDGE_SCRIPTS_DIR "pane-status.ps1")
@@ -342,7 +344,6 @@ function Remove-ProfileExcludedSupportScripts {
                 "orchestra-layout.ps1",
                 "orchestra-ui-attach.ps1",
                 "pane-control.ps1",
-                "pane-dispatch-detect.ps1",
                 "pane-env.ps1",
                 "pane-border.ps1",
                 "pane-status.ps1",
